Move Module code out of the Model and into a place that feels appropriate LibreNMS\Modules
