+++ html/includes/print-menubar.php
@@ -1,8 +1,8 @@
 <?php
 // FIXME - this could do with some performance improvements, i think. possible rearranging some tables and setting flags at poller time (nothing changes outside of then anyways)
 
-use App\Models\WirelessSensor;
 use LibreNMS\Authentication\Auth;
+use LibreNMS\Modules\Wireless;
 use LibreNMS\ObjectCache;
 
 $service_status   = get_service_status();
@@ -455,7 +455,7 @@ if ($toner) {
         </li>
 <?php
 
-$valid_wireless_types = WirelessSensor::getTypes(true);
+$valid_wireless_types = Wireless::getTypes(true);
 
 if (!empty($valid_wireless_types)) {
     echo '<li class="dropdown">
